Code refactor: extracting methods to enum classes.
The TaskTypes stores all the enums of the different class types and each class does

own thing so it is not stored all in one method.

Furthermore, the different commands in the parser class are allso in charge of their

own logic. All of the regexing is contained inside of the method and is not done in

main method.

By doing this, the code is more readable and easier to debug also.