Event class: store starting and ending time as a LocalDate
The Event class stores the starting and ending date as strings,
which creates redundancy, especially since there
is a DateConverter class as well.

This redundancy therefore makes working with dates more difficult
and potentially leading for them to be more errors.

Let's store the starting and ending dates directly as LocalDate
objects inside of the Event class and remove the DateConverter
class.

By doing so, you can improve code readability.