For all classes: add more concise Javadoc comments for my methods
The method Javadoc is too vague and therefore makes it difficult
for readers to understand.

Rewriting the Javadoc comments, especially the header comments to
ensure that they are all descriptive and mention what errors are
being thrown from each method.

Let's be more concise when writing Javadoc.

By doing so, it is more clearer for readers when viewing the code.